+++ src/main/java/Duke.java
@@ -1,3 +1,5 @@
+import duke.*;
+
 import java.io.*;
 import java.util.Scanner;
