+++ src/main/java/duke/InvalidInputExpression.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class InvalidInputExpression extends Exception {
     public InvalidInputExpression(String input) {
         super(input);
