+++ src/main/java/duke/Task.java
@@ -1,8 +1,6 @@
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.io.BufferedReader;
-import java.io.FileReader;
-import java.io.FileWriter;
+package duke;
+
+import duke.DukeException;
 
 public class Task {
     String list;
