+++ src/main/java/duke/TaskList.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.*;
+
 import java.io.IOException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
@@ -217,10 +221,10 @@ public class TaskList {
     }
 
     /**
-     * Adds a ToDo task to the task list based on the provided input.
+     * Adds a duke.ToDo task to the task list based on the provided input.
      *
      * @param input The user input containing the task description.
-     * @return A message indicating the success of adding the ToDo task.
+     * @return A message indicating the success of adding the duke.ToDo task.
      * @throws DukeException If there's an issue with the input or task description.
      */
     public String handleTodoTask(String input, String from) throws DukeException, IOException {
@@ -250,10 +254,10 @@ public class TaskList {
     }
 
     /**
-     * Adds a Deadline task to the task list based on the provided input.
+     * Adds a duke.Deadline task to the task list based on the provided input.
      *
      * @param input The user input containing the task description and deadline.
-     * @return A message indicating the success of adding the Deadline task.
+     * @return A message indicating the success of adding the duke.Deadline task.
      * @throws DukeException If there's an issue with the input, task description, or deadline.
      */
     public String handleDeadlineTask(String input,String from) throws DukeException, IOException {
@@ -326,10 +330,10 @@ public class TaskList {
     }
 
     /**
-     * Adds an Event task to the task list based on the provided input.
+     * Adds an duke.Event task to the task list based on the provided input.
      *
      * @param input The user input containing the task description and event timings.
-     * @return A message indicating the success of adding the Event task.
+     * @return A message indicating the success of adding the duke.Event task.
      * @throws DukeException If there's an issue with the input, task description, or event timings.
      */
     public String handleEventTask(String input,String from) throws DukeException, IOException {
