+++ src/main/java/duke/Ui.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class Ui {
 
     String divider = "------------------------------------\n";
